ARCHIVE: Improve retry functionality and performance
Improve the performance in case of a retry by making the difference
between transfer errors i.e. file does not exist or has wrong size
or checksum value and metadata errors i.e. the owner or the permissions
are not the correct ones. By doing this we can now skip retransferring
successful files in case of a retry.