CMAKE: Make cmake aware of git tags and generate the version info automatically
CMake is now able to pick up automatically the version of the packages that
it needs to build based on the last git commit. If the commit it tagged then
the version information is extracted from the tag, if not the version info
is extracted using the date of the commit and the hash value of the commit.
This improvement will help to better integrate EOS with the Jenkins build
system and the continuous integration YUM repositories. If the user wants
to set a particular version he can still do this by providing custom values
for VERSION_MAJOR, VERSION_MINOR, VERSION_PATCH and RELEASE when running
cmake using the "-DVAR=VALUE" option.

Conflicts:

	CMakeLists.txt
	cmake/EosUtils.cmake