BACKUP: Enforce layout and checksum for version file and skip checking         atomic version files.
For files starting with ".sys.v#." we enforce a 2-replica layout and
we also set explicitly the checksum when copying the file. This ensures
the verification step is done successfully since otherwise we don't
have any checksum type and value saved for these type of files. This
in it's turn is because the directories containing these files have no
exteneded attributes.

We also skip atomic version files for the time being i.e. files
containing ".sys.a#.v#".