DOC: Refactor the generation of the documentation
Add dependency on Sphinx for documentation generation. The documentation
can be generated by using the make target "doc" which in its turn calls
the "generate_docs.py" script. This script is responsible for generating
the CLI documentation by formating the output of each eos command called
with the "-h" flag and also builds the rest of the .rst documentation that
resides in the /eos/doc file.