MGM/NS: Move the namespace quota accounting into Redis
This involved modifying the IQuota interface to implement the necessary
methods to do various operations efficiently.
Update the MGM quota implementation to use a new interface - but
the changes were minimal in this part of the code.
Implement the namespace quota classes that talk directly to the
Redis instance - the bulk of the commit.