NS: Add unlinkFile method to the IView interface that takes an IFileMD object as parameter. This makes the clean-up process much simpler.
Add posibility to configure the Redox client at each view level so that testing
is much simpler and can be easily configured to use the desired Redis instance.

Add clean-up of fsids which are not used any longer i.e. they don't hold any files

Add protection when removing a container to first check that it really is empty.