CMAKE: Modify the convention of package versioning see www.semver.org
A tagged release will have the following format:
MAJOR.MINOR.PATCH-RELEASE i.e. 4.0.11-1
A commmit release will have the following format:
MAJOR.MINOR.PATCH-RELEASE where the first three values will be that
of the most recent commit on the current branch and the RELEASE will
have the following format:
devYYYYMMDDgitxxxxxxx the last 7 characters represent the hash of the
current commit from which the release is built.