NS: Mark a file as inconsistent as soon as any notifications are sent to the listeners.
Reduce as much as possible the FileMD interface - namely remove the
SetConsistent and IsConsistent methods.

When a file is removed from the namespace then force the flush of
consistent files to the backed so that if a crash happens at this
point we are not left with a fid to check that doesn't correspond
to a file object.