MGM: Fix wrong permissions on the "conversions" and "archive" directories when a new EOS instance is booted.
The "conversions" and "archive" directories were owned by root when a new
EOS instance was creted and then on a subsequent restart of the MGM the
ownership was transferred to daemon. This lead to "Operation not permitted"
errors when the Converter was trying to run - issue reported by the Comtrade
team.