FST: Fix removal of blockxs files
The blockxs files corresponding to data files were never deleted since the
deletion was done through the FileIo object of the data file. The FileIo
has no knowledge of the blockxs file. To have this working for local files,
one has to do the deletion through the XrdOfs::rem interface which in its
turn will call the XrdOss interface and this properly cleans up the data
and the blockxs files.