MQ: Refactor the Set method and remove the ModificationTempSubjects map
Remove the option of doing a Set operation without taking all the locks
inside the method implementation. Also drop the use of the
ModificationTempSubjects as there was no added value by collecting all
the notifications and doing them one by one. This is now done as usual
inside the Set method.