NS: Remove locations while iterating, instead of all at the end
The LocationRemoved change listener in FileSystemView relies on
having reliable values for getNumUnlinkedLocation and getNumLocation.

This means that if we clear the locations all at once in the end,
the listener has no chance to know if the replica count of a file
has reached zero, and thus correctly place it in no-replicas-set.