MGM: Refactor the proc interface to allow greater flexibility and encapsulation. Partially implement server side ACL command interface.
- Implement the Factory design pattern in the ProcInterface class for creating
specific ProcCommand object for dealing with different types of commands.

- Create IProcCommand interface which should be implemented by both old style
commands passing all the information as opaque info in the URL and new style
commands that use ProtocolBuffer objects which are serialized and base64encoded.

- Implement the AclCmd on the server side as a possible reference implementation
that can deal with long running tasks on the server side and not trigger client
retries in the meantime.

- Add neccessary functionality for storing all uid/gid(s) in the ACLs in numeric
format and to convert them to string representation when the acl command is used
by the client. If one wants to see the numeric representation then he can use
the "attr ls user.acl/sys.acl" command.

- Separate the ProcInterface and ProcCommand classes in their own header and
source files.