FUSEX: Move XrdClProxy tests into separate test executable
Let's follow the convention used in QuarkDB in here, as well.
eos-fusex-tests are the "fast" tests, meant to be ran together
with make, ie "make && ./tests/eos-fusex-tests" and should not
take more than a couple of seconds to run.

eos-fusex-stress-tests are for longer-running tests, those
which would be too annoying to have to run alongside "make".