NS: Add FileIterator interface for going through the file ids on a particular filesystem rather than copying the entrie list.
Add hasFileId method to check if a file id is stored on a particular file
system.
Add getNumFilesOnFs/getNumUnlinkedFilesOnFs to count the number of files/
unlinked files on a particular file system. There are several places in the
code where we only need this type of information rather than the entire list
of files. This should considerably improve the performance and memory usage
of those sections of code.