FUSEX: Take into account child/parent relation for rename lock-order
Just using increasing inode number is not enough, since in all
the rest of the code we first lock the child, and then the parent.

If the two directories are related in this way (child and parent),
we _have_ to follow the above convention, or we risk deadlocking.

Otherwise, we decide based on increasing inode.

After running many times with tsan I got no lock-order-inversion
warnings, so we're probably good.