NS: Use collection of mutexes to sync the retrieving of container md
This is useful in case the container contents is very big or the backend
is slow. In such cases we might get multiple requests for the same
container md and we end up retrieving the same data multiple times. This
is not a problem from a functionality point of view but it's just
inefficient use of I/O and network towards the backend.