FST: Add fst.valid tag to avoid race conditions triggered by open replays
The fst entry server will set the fst.valid tag to a point in time equal
to the current timestamp + the configured XRD_STREAMTIMEOUT value. In this
way we are protected from race conditions triggered by slow open procedures
on the other replica/stripe(for rain) files which determine the XrdCl client
to retry the inital open operation.

This race condition was triggering two close event and a commit with 0 size
was sent to the MGM. The first entry point always had the correct size and
checksum but the second replica had 0 size.