MGM: Avoid corruption of the configuration file when MGM is restarted in a loop. Fixes EOS-2249
The corruption happened because the config engine was registered with the FsView
before the Master was loading the actual configuration from disk. If in the meantime
the MGM crashed and there was any global udpated done though the FsView object that
would require saving the config - then we would end up with a configuration on disk
that only had the last update.