MGM: Advertise the MGM as TPC capable to allow transfers of 0-size files. Fixes EOS-2442
For 0-size files the MGM will not redirect the client to an FST and for
such cases when a TPC transfer is required the transfer will fail. The MGM
is queried by the XrdCl client so see if it supports TPC transfers and this
is enabled by setting the XRDTPC env variable.