MGM: Avoid closing the file descriptor managed by the XrdOucStream twice
The value of the file descriptor used for reading in the xrd.cf.mgm config file
was closed twice. The problem was that in the meantime a new file was opened and
the file descriptor was assigned to this new file (in this case the first fan out
log dir i.e. Balancer.log) and later on this file descriptor was closed again.
During the clean up, I've also removed the autorization plugin which is not used.