NS: Drop support for transactional recursive delete
Supporting transactions in the namespace is most
likely out of the question, as POSIX itself doesn't,
and could hugely complicate a lot of things.

Having a transactional, atomic recursive delete in
the core namespace code is not a good idea, and would
be very difficult to accommodate once we introduce
fine-grained locking.

Happily, none of the MGM code uses it.