NS: Ensure writes to fs handler during cache load are reflected in resulting contents
There's a race condition between triggering a cache load, and
receiving the contents.

We don't want to lock the contents during this window, as we
might block writers this way.

We record all writes to the fs handler into a ChangeList,
and replay that once the contents have arrived.

The ChangeList is only active while cache loading is in
progress.