stagerrm now calls OFS->_dropallstripes() with forceRemove set to false
Setting forceRemove to true was incorrect because it forced the
EOS namespace to forget the physical files in the FSTs.

Setting forceRemove to false is correct because it adds the file
to be deleted to a list of files to be unlinked on their local
FST storage.

Thanks to Jozsef for spotting that stagerrm needed fixing.