COMMON/UNIT_TESTS: Add XrdConnPool class for managing XRootD physical connections
This class can now be used throughout the code especially on the MGM side when
doing drain operations.

The similar functionality in fst/io/xrd/XrdIo.hh can be replaced with this class.

Added unit tests for XrdConnPool and also for the helper class XrdConnIdHelper
which ensures automatic assignment and relase of the connection id.