FST: Encapsulate requests for broadcast in a separate method and use the non-blocking method to retrieve the FST config queue in the Communicator thread to avoid a deadlock.
The deadlock might appear when the MGM sends updates about the FST config but not about
the FST config queue and then it blocks waiting for a value that only he can set. Remove
also the "random" sleep from the start-up of the FST which was a "workaround" for this
issue.