[MGM] Don't return SFS_ERROR from XrdMgmOfs::prepare
A prepare stage request contains multiple files. The prepare request
can succeed for some files and fail for others. The old behaviour was
to stop on an error and return the error code. However, some files may
have been staged but the client has no request id to track it.

The new behaviour is always to return success (SFS_OK or SFS_DATA) and
to process all files in the list, even if some of them synchronously
fail. In this way, the client always gets the request id and all files
that can be staged will be.

It is up to the client to check the status of each file to see if the
request succeded, failed or is still in-flight, but this was anyway the
case, as a later asynchronous failure is possible even if the original
synchronous request succeeds.