mgm: FuseServer/Caps: replace erase calls with find
Since we are planning on erasing the found caps, use find so that we can reuse
the found iterator to erase which gives a slight O(logn) advantage instead of
using the operator[] on maps again. Also reuse the return value of erase in
`Remove()` call as this already indicates whether an erase was done on a key or
not.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>