tests: add a sample google benchmark test
This uses the system package right now as a submodule currently needs to toggle
the EOS CXX build flags as google/benchmark brings its own #define for version
which causes build failures. The options for workaround include using External
project or toggling the flags for a subdirectory. We've moved the tests in
test/gbench so as to not conflict with our own benchmark implementation

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>