mgm: FuseServer: Caps: move to unordered variants of the containers
This allows for a future change where we rely on hash tables which don't promise
the ordering of keys like a std::map does.

Also GetCaps() returns a const. ref to the caps table rather than a mutable
reference since the only user of the function seems to only require a view
rather than a modify

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>