unit_tests: add basic cap testing functionality
This introduces a basic couple of tests to test the base cap functionality. The
idea is to test for this in a multi threaded way to ensure that we provide the
consistency guarantees and we don't break anything when we change the underlying
hashmaps etc.

We currently still create an instance of XrdMgmOfs, with some of the services
set to 0, a TODO item is mocking this as much as possible. The common XrdMgmOfs
is shared across test suites so as to prevent costly construction/destruction
per test suite

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>