common: Introduce a concurrent_map_adapter
This data structure basically internally holds a given hashtable type and adds
concurrency primitives to it. Different Lock types are also supported as far as
they can have a UniqueLock & SharedLock type to provide for reader writer
primitives. While it is intended to be a close replacement to the standard,
deviations are expected in apis returning mutable references and pointers as
modifications outside of the map will lose the internal locking.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>