common: concurrent_map: implement erase(iter)
This function currently returns an iterator to the next element if the
underlying map has such a return type. This might change in the future where a
void return might make life easier if using other maps like google::densemap
where an iterator is not returned. The postfix++ iterator is also implemented
though very much wip.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>