mgm: XrdMgmOfsDirectory: take allowals & denials of ACLs
Regardless of mod permissions, allowals & denials of ACLs will re-grant
or deny access to listing a folder. This change builds on the previous
behaviour where an ACL grant will override an allow to also similarly
honor a deny. Since we have to always check for a deny, we have to read
the ACLs always unlike the previous case where we checked for a deny
only iff mod permissions are -ve

Signed-Off-By: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>