common: Mapping: use static strings for known constant keys
Otherwise C++ creates a std::string every single time. On a microbenchmark the
static string version was 80% faster than the version creating a std::string
from const char*. A decent middle ground in the future if we do this a lot could
be to use transparent key comparision where a std::string_view can also be used.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>