MGM: Handle the case for fsck repair of RAIN files when there is a rep_diff_n error
In this case there is no stripe that needs to be replace but new stripes need to be
added. In order to use the existing functionality of the DrainTransferJob code, we
add the sentinal file system 0 to mark the file system to be drained/repaired. In
this way no exsting(valid) stripes are dropped but only missing stripes are
reconstructed in place.