test: EosHashBenchmark: implement single run mode & random inserts
This allows the hashbenchmark to take random insertion order(r) as well as run all
the tests together for one hashmap(s), releasing the memory, so larger map sizes
can be tested. A special flag is added to skip ulib map as insertions go to a
crawl after 500 mill keys when random insertion is attempted.

Fixes: EOS-4863
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>