common: Introduce an Iterator version of String Splitter
A new non owning (by default) string splitter providing an iterator like
interface is introduced. std::string_view is the default type of input & output
however this is customizable, the delimeter also can be made use a char which is
helpful for single character splitters, mostly for null characters. Performance
wise this version seems to use about 30% of the time compared to a std::string
vector splitter, and a copy version is still uses only about 40% of the time.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>