common: StringSplit: use a helper function to calculate next position
This helps to change the implementation later in case we need a all_of pattern
match instead of the any_of default we do. For the future if we have such a use
case it would be fairly easy to accomplish this with an additional template
paramter or tag.

Also fix indentation of inner iterator class and add tests for mutli character
splits

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>