fst: kill global FmdDbMapHandler part deux
We move to using gOFS.mFmdHandler wherever possible. To handle call sites using
DB specific calls we introduce `gOFS.FmdOnDb()` which returns true if we are
holding a DbMap handler. Trim is a thread that will only be launched in case we
are storing fmd on DB. Since trim may or may not outlive the original object, we
actually hold the shared_ptr itself unlike other cases where we do a normal
downcast on the raw ptr.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>