mgm: GroupBalancer: move Balancer Engine into its own component
This introduces the following changes:
- create a BalancerEngine base class
The main purpose of the engine is to pick & classify groups based on some
configurable. The common functionalities are added into a base class. Also add a configure
method so that child classes can configure themselves accordingly since this is
the format we receive the data anyway.
- Move current balancer into StdDevBalancerEngine
The core logic is moved into this, add gtests to demonstrate behaviour
- Introduce a BalanceInfo Fetcher interface
While we could potentially drop the interface and make this just a regular
function, it is useful for Dependency Injection if we manage to mock some of the
FsView interfaces
- use sets for over/under avg group info
Since we don't seem to use the ptr to Group for almost all cases, just store the
names only. Since it is easy to convert the name to a FsView::Group just convert
this at the callsite. This should make the core of GroupBalancer easier to
separate from the part that deals with MGM/FsView.
- make the GroupSize map transparent
This allows for std::string_view/std::string comparision of keys when needed.

Fixes: EOS-5067
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>