mgm: GroupBalancer: drop record transfer on file move
Since the move of file is only queued, the actual stats aren't yet updated. Also
since we only store the avg total used size & capacity this calculation was
wrong. While it is possible to actually compute the total and store this as well
in the GroupSizes struct, it is mostly wasteful as by the next expiry of cache,
this value would have been reset to the FsView's actual numbers anyway. If this
turns out to be really an issue we can consider either holding the total and/or
reducing the cache times.

Fixes: EOS-5079
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>