mgm: GroupBalancer: sync. internal groupbalancer map access
While GroupBalancer itself is single threaded, so various internal engine
maps/sets need not be locked, the Status call will be from the cli and thus a
different thread. So protect the map with a mutex which will only be used in the
populate Call as this is the only write call. Since the groupbalancer itself
just runs `GroupBalance` forever and calls nothing in another thread, no other
reads need this lock.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>