mgm: GroupBalancer drop recalculate
Originally we calculated and recomputed averages every cycle and updated these
numbers while scheduleTransfer was called. As a fix for EOS-5079, this was
dropped as these computations were incorrect as we subtracted actual size from
an average, however even if we subtract against a total size this number would
not be realized until Converter finishes the transfer, so the next populate call
will reset the info to something different from our computed numbers anyway.
Drop recalculate as this currently does no re classification of groups.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>