MGM: GroupBalancer: reconfigure only on demand
We set an atomic needs_reconfigure from SpaceCmd only when we encounter a
groupbalancer related command. We use acquire/release barriers on the atomic,
while even relaxed semantics should work, in case of release from SpaceCmd we
know that all the memory writes are commited so the unlock mutex will not
contend from GroupBalancer.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>