mgm: GroupBalancer: protect group picker with a rlock
This protects the access to the group itself, the routines chooseFileFromGroup &
transfer again grab their own locks, we do crosscheck for nullptr in any case so
this should almost cover all the bases. In case a there is a write operation in
between these function calls during which we're not locked is the only scenario
that might hit us. The only way to prevent that would be to recheck the Viewmap
and obtain the group locally in every function call. However we do obtain other
file ptrs from QDB in these calls and they should return null preventing this in
almost all cases.

Co-Authored-By: Manuel Reis <manuel.b.reis@cern.ch>
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>