TESTS: add basic test for scandir config
ScanDir source is explicitly added to FST UT sources to ensure that we pass on
the NOOFS flag. Also, since we compile the ScanDir as a part of unit-tests-fst
and the underlying libraries do not expose XXHASH directly, add a link.
Additionally since xxhash assumes common path, provide a target include on
common.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>