MGM: GroupsInfoFetcher: internally store the filter function
This drops the dependency of any filter function requiring to inherit from
GroupStatusFilter, we internally store the function object, a constructor is
provided with a template argument allowing for any callable type that can
implement a operator()(status) argument. This is similar to std::function but a
bit more flexible in accepting lambdas and such.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>