MGM: GroupDrainer: mark completed groups as drained/drainfailed
This adds a new status string drained/drainfailed to groups and
sets them if we identify all the FSes in the group are drained. We
reuse the GroupStatus enum from earlier instead of a GroupDrainStatus.
This makes it easier to reuse this function to set a generic groupstatus,
however checkGroupDrainStatus function now returns an enum that isn't
exactly mapping to an actual groupstatus but maps well with a groupdrain status.
This direction was chosen for now in favour of having 2 enums with the same
content and similar enum transform functions. Also check the status in
groupdrainer-tests which has a reduced retry_count in the CI to save some time.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>