MGM: GroupDrainer: fix an issue that first groups/fses were always picked
Once we reach full amount of transfers, and converter starts clearing up a
few transfers, we reach this position where we only are left with a few
slots remaining, since we always loop from index 0, this meant that the index
0 group and FS were almost always favoured inspite of other sources and targets
being available. Avoid this by using a global uint16_t counter that'll be used
as a seed for the RR picker.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>