fst: FmdHandler: Introduce FmdHandler class & ensure RAIN Layouts
this commit introduces FmdHandler, which basically implements all the non Writer
specific implementation. The derived classes only need to implement the
Put/Get/Commit functionality and the rest of the common logic can be reused.

FST: Ensure for the RAIN layouts that the reference size in LevelDB matches the
logical size of the file and not the physical size of the local stripe. This is
fine for replica layouts but leads to HTTP errors while reading RAIN layouts.

(cherry picked from commit 74baf636dc7722fae344dbbf58109411e0b477da)
the Update With Disk/MGM functions were updated manually to match the FmdDbMap
changes

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>

Conflicts:
	fst/filemd/FmdDbMap.cc: UpdateWithMgmInfo conflict after msg rename