MGM: GroupDrainer: use a seed per group for picking FSes
For certain values of total groups and FS per group, the current method
will end up skipping over a few indices esp. if the amount of groups draining is
even; avoid this by using a per group counter for picking FSes so we replicate
the group picking logic. Simulated tests are added that show this behaviour.

Fixes: EOS-5269
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>