COMMON: Introduce a simple ConfigStore interface
This is a simple interface that allows to save/load a key/value from some config
store provider. The idea is to abstract away whether we talk to a global config
store/per space level/qdb level config store. This should be implemented by the
subclasses so it is easier for future dependency injection like scenarios where
we can test some aspects without needing to have a qdb/mq etc present.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>