MGM: Allow Transfers Queue to be reset & protect all mMaxTransfers access
reset command now will also reset all transfers queue. Additionally we protect
all accesses of mMaxTransfers as well using the same mTransfersMtx to ensure
that we don't have incorrect reading of sizes. Status output now includes the
currently configured value of max transfers so that it is easy to judge if the
config value from the space is reflected in our internal logic.

Fixes: EOS-5285
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>