FST: Introduce a utility class to Get FSID/File paths
This class basically provides 2 functions to get FSID path and File Path, the
idea would be that interfaces that take a path for eg. in FmdAttr, which can
be mapped back from files. The idea with this is that we can remove the
dependancy on XrdFstOfs for standalone tools since we'll hold a ptr to the
Utility class which can be constructed with or without XrdFstOfs depending on
the arguments supplied.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>