MGM: GroupDrainer: track scheduled transfers to avoid reschedules
In case of erasure layouts since listing the file list per FST may include files
already scheduled files (since we pop off the successful transfers, and the list
of file MD might be stale by the time converter schedules them), track all
scheduled transfers and avoid these. We drop this tracked Transfers when we
reset failed or all transfers. A future improvement could the be the use of a
bloom filter in this as we only care about files not present in a set.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>