MGM: Make sure there are no data races/leaks when initializing the ldap connection. Fixes EOS-5308 According to the man page of ldap_initialize:
Note: the first call into the LDAP library also initializes the global options for
the library. As such the first call should be single-threaded or otherwise protected
to insure that only one call is active. It is recommended that ldap_get_option() or
ldap_set_option() be used in the program's main thread before any additional threads
are created.