COMMON: Observer: fix potential bug when move only arguments are involved
With univ. references, when a Observer is passed on temporary values this may
decay to a const ref & rvalue type even though the intended signature was a
value type. Since we explicitly type the value types when creating the
ObserverMgr itself, use this info when creating the packaged task instead of
forwarding as otherwise with moveable types or temporaries we'll end up moving
the argument in the first instance which is undesirable. A test is added to
catch this behaviour using a string type, and we validate that we've valid
strings passed on to multiple observers which accept different variations of
std::string

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>