FST: FmdConverter: improvements in ConvertFS routine
Avoids figuring out the fsid every time. Also make sure that this is a supplied
argument so that we don't have to parse it for each file. 0 fids are skipped and
the last function takes ownership, so that we'll not have to deal with outlived
references as the thread it executes in may be different from the calling thread

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>