MGM: Policy overhaul
+ Support for a is_local opaque flag wherein we don't look for
persistent/transient keys. Space cmds also support a "local.policy" alongiwth
"policy." settings. This is an optional setting that is only enabled when opaque
flag localconfig=1 is set atm.

+ Use batch apis client keys or from qdb. We use Batch APIs to get keys and
populate keys with a helper function that walks through the priority of app >
user > group > space for RW keys.

+ Avoid double lookups where we use the map's value later.
Also use constants in code, this avoids potentially constructing a string while
doing a map lookup, while C++17 offers transparent lookups where keys are
guaranteed not to be constructed, this requires the map itself to be defined
with a std::less<> which requires changes in the API, however we can still
already construct these strings beforehand.

+ MGM: Policy: use XrdUtils::GetEnv for easier value fetching

+ MGM: Policy simplify generating the various RW policy params

A simple RWParams struct holds this info and can translate/create the various
policy RW strings as needed.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>