MGM: Policy overhaul (╯°□°)╯︵ ┻━┻
Support for a is_local opaque flag wherein we don't look for
persistent/transient keys. Space cmds also support a "local.policy" alongiwth
"policy." settings. This is an optional setting that is only enabled when opaque
flag localconfig=1 is set atm.

The rest of GetSpacePolicyLayout however is independant of whether we get local
client keys or from qdb. We use Batch APIs to get keys and populate keys with a
helper function that walks through the priority of app > user > group > space
for RW keys.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>