NS: Added a lock registration mechanism and a locking class to the IFileMD object
The locker class has to be used as an RAII class. The instanciation of it will
lock the underlying IFileMD pointer and register the lock.
The destruction of the locker class will unlock the IFileMD and unregister the lock.

This class can be used for doing multiple operations on the underlying IFileMD object
under the same lock.