COMMON: Introduce an ExecutorMgr for folly or eos threadpool executions
This is a class that holds either an eos::common::Threadpool or a folly executor
and allows scheduling functions in these. Since the APIs are vastly different
on adding a function to the executor/threadpool ie. the callable type expected
is a std::function vs a folly::function which are not exactly interopeable, this
class is a bridge to unify these intefaces. A free function execVia is
introduced which can execute a given callable either inside a threadpool or a
folly executor where we start with a Future::Unit and chain the function in the
then which executes in the executor. In the ExecutorMgr we just hold a variant
of two types and do classical visitor pattern to dispatch the right Add call.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>