FST: FmdConverter: add support for threadpool as a execution target
Conversions can be run either in folly::executor or a eos::common::Threadpool
depending on what the FmdConverter is constructed with at runtime. We default
to threadpool atm.

In ConvertFS routine we hold the results in an opaque future object and wait
until they are realized (effectively calling future.get()). A --executor flag
is also added in eos-filemd-convert utility to test these.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>