FST: Storage: create the executors for Converter at startup
This avoids potential lifetime issues seen in both Threadpool and folly
executors which seems to have issues when the unique_ptr holding converter is
reset and a new instance of the threadpool is created. Instead pre create the
threadpool at storage start, so that this outlives FmdConverter. Allows
FmdConverter not to be created as a ptr anymore as this doesn't really own
anything.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>