COMMON: ShardedCache: provide option to initialize after construction
This creates a ShardedCache with no threads and no mutexes when default
constructed, this is mainly useful for static longlived caches that may not be
needed everywhere which is the case with current Mapping. Since we only need the
ShardedCache in MGM only and not the rest of the components. A future plan could
be to use mgr class to manage these caches and drop this 2 phase initialization.

In FUSEX for eg: we'd fork after loading, which would mean the static maps would
have threads in an undefined state

Fixes: EOS-5442
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>