COMMON: ShardedCache: add option to construct without GC
ShardedCache has a constructor without a ttl that would just function as a
sharded map without running the GC thread. This is useful over providing an
invalid default constructor as we'd still be able to use the sharded map as such
without using the GC features. Also helps for 2 phase initialization for static
maps wherein we'd still start in a valid state, just that GC threads can be
explicitly started via reset_cleanup_thread.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>