COMMON: ThreadAssistant: introduce a setSelfThreadName method
since setting thread names on other threads not spawned by the main thread can
potentially fail depending on permission issues, add a small helper method
to set self name which should usually succeed. Since ThreadAssistant is
the one coordinating the long running loop, this makes it easy to set this as
the first call before we run the long running loop. The other alternative is
from the AssistedThread to actaully do this in the main thread call itself
by decomposing the args correctly, which might be a bit more involved.

Fixes: EOS-5446
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>