COMMON/MGM: Mapping: use the sharded variants of Uid/Tident cache
The ActiveUids & Tidents are moved to the sharded variants. The who command
basically copies the entire sharded map, however this will not extend the
lifetime of existing entries and is shorter wrt locks on IdMap. A future TODO is
to remove the existing Active* maps once we're confident with this

Fixes: EOS-5459
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>