MGM: placement: Introduce a simple flat scheduler
This scheduler basically schedules assuming a simple flat array of storage
elements and disks supplied as cluster_data. The elements are chosen by a
PlacementStrategy class, currently a simple Round Robin scheduler is introduced.

The rule to pick elements right now would be a simple array indexed by the
storage element type and the elements to pick. A -1 value implies pick as many
replicas as originally requested, this allows for a more storage aware placement
strategy where one can tweak the numbers, so 2 replicas in site 1 and 1 replica
in site 2 split across two racks for eg.

For the default case, where we don't expect geo scheduling, we do a kind of DFS
to go to the end of the tree, since we are only about to pick 1 site, group etc,
this makes it simple.

The non default case picks the replicas using a classical BFS approach of
picking elements per storage element

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>