MGM: placement/scheduler: Use a circular fixed size vector for epochs
We store the last 50 epochs, as shared_ptrs so that threads already using the
shared_ptrs are not affected. Protect the multiple writer case with a mutex,
this could be achieved by a semaphore/ writer spin lock also as we don't
envision multiple threads parallely adding groups and disks. Readers are still
lock free and wait free, as we don't allow modifications to the structure of
ClusterData, only the atomic statuses.

The whole fixed circular buffer of shared_ptrs might be moved to its own
structure as this is not really dependant on the core ClusterScheduler logic

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>