FST: Fix 0 size files being marked as rep_missing_n and reorder some FST operations for better consistency. Fixes EOS-5498
For the fmd as xattr the FST was fist trying to attach the fmd entry but
for new file there is no file on disk, and the CreateFile method would
fail for paths that also required creating the parent directory. Therefore
all these entries were having the rep_missing_n flag set. This is now
changed so that attaching the fmd xattr or the db entry are create only
after the actual open of the file on disk is successful.